refactor: move countdown and sponsors components to a new main directory for better organization and maintainability
